feat: separate state management logics and component
